refactor(web): extract v1 run creation helper to api-test-helpers
Move the local helper function from route.test.ts to api-test-helpers.ts
for reuse consistency with createTestRun.
