+++ turbo/apps/web/app/v1/runs/__tests__/route.test.ts
@@ -8,6 +8,7 @@ import { randomUUID } from "crypto";
 import {
   createTestRequest,
   createTestCompose,
+  createTestV1Run,
   completeTestRun,
   createTestModelProvider,
 } from "../../../../src/__tests__/api-test-helpers";
@@ -24,28 +25,6 @@ vi.mock("@aws-sdk/s3-request-presigner");
 vi.mock("@axiomhq/js");
 
 const context = testContext();
-
-/**
- * Helper to create a run via v1 API and return the run ID.
- */
-async function createV1Run(
-  agentId: string,
-  prompt: string,
-): Promise<{ id: string; status: string }> {
-  const request = createTestRequest("http://localhost:3000/v1/runs", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ agentId, prompt }),
-  });
-  const response = await createRun(request);
-  const data = await response.json();
-  if (!response.ok) {
-    throw new Error(
-      `Failed to create run: ${data.error?.message || response.status}`,
-    );
-  }
-  return data;
-}
 
 describe("Public API v1 - Runs Endpoints", () => {
   let user: UserContext;
@@ -66,7 +45,7 @@ describe("Public API v1 - Runs Endpoints", () => {
     testAgentId = composeId;
 
     // Create a completed test run for read operations
-    const run = await createV1Run(testAgentId, "Test prompt for runs API");
+    const run = await createTestV1Run(testAgentId, "Test prompt for runs API");
     testRunId = run.id;
 
     // Complete the run so it doesn't block concurrent limit
@@ -168,7 +147,10 @@ describe("Public API v1 - Runs Endpoints", () => {
 
       // Create run as other user
       mockClerk({ userId: otherUser.userId });
-      const otherRun = await createV1Run(otherComposeId, "Other user prompt");
+      const otherRun = await createTestV1Run(
+        otherComposeId,
+        "Other user prompt",
+      );
       await completeTestRun(otherUser.userId, otherRun.id);
 
       // Switch back to original user and try to access other user's run
@@ -188,7 +170,7 @@ describe("Public API v1 - Runs Endpoints", () => {
   describe("POST /v1/runs/:id/cancel - Cancel Run", () => {
     it("should cancel a pending run", async () => {
       // Create a new run (starts as running)
-      const run = await createV1Run(testAgentId, "Run to cancel");
+      const run = await createTestV1Run(testAgentId, "Run to cancel");
 
       const request = createTestRequest(
         `http://localhost:3000/v1/runs/${run.id}/cancel`,
@@ -402,7 +384,7 @@ describe("Public API v1 - Runs Endpoints", () => {
 
       try {
         // First run should succeed (creates running run)
-        const run1 = await createV1Run(testAgentId, "First concurrent run");
+        const run1 = await createTestV1Run(testAgentId, "First concurrent run");
         expect(run1.status).toBe("running");
 
         // Second run should fail with 429
@@ -432,9 +414,9 @@ describe("Public API v1 - Runs Endpoints", () => {
 
       try {
         // Create multiple runs - all should succeed
-        const run1 = await createV1Run(testAgentId, "Run 1 with no limit");
-        const run2 = await createV1Run(testAgentId, "Run 2 with no limit");
-        const run3 = await createV1Run(testAgentId, "Run 3 with no limit");
+        const run1 = await createTestV1Run(testAgentId, "Run 1 with no limit");
+        const run2 = await createTestV1Run(testAgentId, "Run 2 with no limit");
+        const run3 = await createTestV1Run(testAgentId, "Run 3 with no limit");
 
         expect(run1.status).toBe("running");
         expect(run2.status).toBe("running");
